==> P7_SVM_face.ipynb <==
import pandas as pd
import numpy as np
from matplotlib import pyplot as plt
import cv2
import os
import re
from sklearn.model_selection import cross_val_score  #交叉验证
from sklearn.svm import SVC  #支持向量机
from sklearn.metrics import accuracy_score  #混淆矩阵准确度
from sklearn.metrics import classification_report  #混淆矩阵分类性能评估报告
from sklearn.metrics import confusion_matrix    #混淆矩阵数组

## 说明：
本程序为使用dlib库128维特征距离数据集进行SVM情感分类的简单实验

trainDistance = np.load("data/trainDistance.npy")
testDistance = np.load("data/testDistance.npy")
DS_face_map = np.load("data/DS_face_map.npy")
DR_face_map = np.load("data/DR_face_map.npy")
DS_face = np.load("data/DS_face.npy",allow_pickle=True)
DR_face = np.load("data/DR_face.npy",allow_pickle=True)

print("DR_face_map: ", DR_face_map)
print("DS_face_map: ", DS_face_map)
print("DR_face: ", DR_face)
print("DS_face: ", DS_face)

不处理

#采用sklearn.svm.SVC，类似的有sklearn.svm.LinearSVC

clf_ovoLinear = SVC(kernel='rbf')
clf_ovoLinear.fit(trainDistance, DR_face)

ovoLinear_face_pre = clf_ovoLinear.predict(testDistance)
ovoLinear_accuracy = accuracy_score(DS_face, ovoLinear_face_pre)
print('ovoLinear_accuracy = ', ovoLinear_accuracy)
print('分类结果报告：', '\n', classification_report(DS_face, ovoLinear_face_pre))

SMOTE过采样

from imblearn.over_sampling import SMOTE
imb_SMOTE = SMOTE(random_state = 0)
trainDistance_SMOTE, DR_face_SMOTE = imb_SMOTE.fit_resample(trainDistance, DR_face)
trainDistance_SMOTE.shape

SMOTE_face_dataframe = pd.DataFrame(data =  {'face':DR_face_SMOTE})
print(SMOTE_face_dataframe)
SMOTE_face_count = SMOTE_face_dataframe['face'].value_counts()
print(SMOTE_face_count)

#采用sklearn.svm.SVC，类似的有sklearn.svm.LinearSVC

clf_ovoLinear_SMOTE = SVC(kernel='linear')
clf_ovoLinear_SMOTE.fit(trainDistance_SMOTE, DR_face_SMOTE)

ovoLinear_SMOTE_face_pre = clf_ovoLinear_SMOTE.predict(testDistance)
ovoLinear_SMOTE_accuracy = accuracy_score(DS_face, ovoLinear_SMOTE_face_pre)
print('ovoLinear_SMOTE_accuracy = ', ovoLinear_SMOTE_accuracy)
print('分类结果报告：', '\n', classification_report(DS_face, ovoLinear_SMOTE_face_pre))
